Create Best NeighborhoodCT
Create a script that uses your own (small) set of criteria to identify the best census tract in Connecticut. A young college educated couple is seeking to move into an apartment in Connecticut. They are looking to rent an apartment priced between the range of $1200 and $1600. They do not like children so don’t want to be near a school and would prefer to live among an older population, with small families and household sizes. For good measure, it would be away from any school.  They also work in Stamford and Norwalk so would prefer to live somewhere with a 30 minute driving distance of both towns or at most a 10 minute drive to one of the train stations.